requirements.txt
multilingual-clip
open_clip_torch
contextlib2
opencv-python
flask
gunicorn
requests
gevent